chore: serial->bigserial id in tables genre and author

+++ spring-05/src/main/resources/schema.sql
@@ -4,13 +4,13 @@ DROP TABLE IF EXISTS genre;
 
 CREATE TABLE author
 (
-    author_id SERIAL PRIMARY KEY,
+    author_id BIGSERIAL PRIMARY KEY,
     name      VARCHAR(50) NOT NULL
 );
 
 CREATE TABLE genre
 (
-    genre_id SERIAL PRIMARY KEY,
+    genre_id BIGSERIAL PRIMARY KEY,
     genre_name    VARCHAR(50) NOT NULL
 );
 
@@ -18,6 +18,6 @@ CREATE TABLE book
 (
     book_id   BIGSERIAL PRIMARY KEY,
     title     VARCHAR(255) NOT NULL,
-    author_id INTEGER REFERENCES author (author_id),
-    genre_id  INTEGER REFERENCES genre (genre_id)
+    author_id BIGINT REFERENCES author (author_id),
+    genre_id  BIGINT REFERENCES genre (genre_id)
 );
